Refactor to own files with export/import
