+++ src/QMap/boot.js
@@ -1,0 +1,45 @@
+import { $dataQMapInfos } from "./constants";
+
+QPlus.request("data/QMap.json")
+  .onSuccess(function (json) {
+    if (json[0] !== "2") {
+      // convert old json type to new
+      if (Utils.isOptionValid("test")) {
+        const fs = require("fs");
+        const path = require("path");
+        const dataPath = path.join(
+          path.dirname(process.mainModule.filename),
+          "data/"
+        );
+        const qMapPath = path.join(dataPath, "QMaps/");
+        if (!fs.existsSync(qMapPath)) {
+          fs.mkdirSync(qMapPath);
+        }
+        const newJson = ["2"];
+        for (let i = 1; i < json.length; i++) {
+          const map = json[i];
+          if (map && map.length > 0) {
+            newJson[i] = true;
+            const filename = "QMap%1.json".format(i.padZero(3));
+            fs.writeFileSync(
+              path.join(qMapPath, filename),
+              JSON.stringify(map)
+            );
+          }
+        }
+        fs.writeFileSync(
+          path.join(dataPath, "QMap.json"),
+          JSON.stringify(newJson)
+        );
+        $dataQMapInfos = newJson;
+      } else {
+        alert("Invalid QMap datatype");
+        window.close();
+      }
+    } else {
+      $dataQMapInfos = json;
+    }
+  })
+  .onError(function () {
+    throw new Error("Failed to load 'data/QMap.json'");
+  });
